docker/ldes-ingest/ldes-ingest-py/pyproject.toml
[tool.poetry]
name = "ldes-ingest-py"
version = "0.1.0"
description = "python ldes ingest client into graphdb"
authors = ["cedricd <cedricdecruw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
apscheduler = "^3.10.4"
pyaml = "^23.9.7"
python-dotenv = "^1.0.0"
sparqlwrapper = "^2.0.0"
pyrdfj2 = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
